utility funcs got moved
