+++ src/app/containers/StoryPromo/LinkContents.jsx
@@ -5,7 +5,7 @@ import { shape } from 'prop-types';
 import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
 
 import { ServiceContext } from '../../contexts/ServiceContext';
-import deepGet from '../../lib/json/deepGet';
+import deepGet from '../../lib/utilities/deepGet';
 import formatDuration from '../../lib/utilities/formatDuration';
 import { storyItem } from '../../models/propTypes/storyItem';
